+++ packages/@markuplint-dev/prettier-config/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.12](https://github.com/markuplint/markuplint/compare/@markuplint-dev/prettier-config@1.0.11...@markuplint-dev/prettier-config@1.0.12) (2025-03-09)
+
+**Note:** Version bump only for package @markuplint-dev/prettier-config
+
+## [1.0.11](https://github.com/markuplint/markuplint/compare/@markuplint-dev/prettier-config@1.0.10...@markuplint-dev/prettier-config@1.0.11) (2025-02-27)
+
+**Note:** Version bump only for package @markuplint-dev/prettier-config
+
+## [1.0.10](https://github.com/markuplint/markuplint/compare/@markuplint-dev/prettier-config@1.0.9...@markuplint-dev/prettier-config@1.0.10) (2025-02-11)
+
+**Note:** Version bump only for package @markuplint-dev/prettier-config
+
+## [1.0.9](https://github.com/markuplint/markuplint/compare/@markuplint-dev/prettier-config@1.0.8...@markuplint-dev/prettier-config@1.0.9) (2025-02-04)
+
+**Note:** Version bump only for package @markuplint-dev/prettier-config
+
+## [1.0.8](https://github.com/markuplint/markuplint/compare/@markuplint-dev/prettier-config@1.0.7...@markuplint-dev/prettier-config@1.0.8) (2024-12-04)
+
+**Note:** Version bump only for package @markuplint-dev/prettier-config
+
+## [1.0.7](https://github.com/markuplint/markuplint/compare/@markuplint-dev/prettier-config@1.0.6...@markuplint-dev/prettier-config@1.0.7) (2024-10-28)
+
+**Note:** Version bump only for package @markuplint-dev/prettier-config
+
+## [1.0.6](https://github.com/markuplint/markuplint/compare/@markuplint-dev/prettier-config@1.0.5...@markuplint-dev/prettier-config@1.0.6) (2024-10-27)
+
+**Note:** Version bump only for package @markuplint-dev/prettier-config
+
+## [1.0.5](https://github.com/markuplint/markuplint/compare/@markuplint-dev/prettier-config@1.0.4...@markuplint-dev/prettier-config@1.0.5) (2024-10-15)
+
+**Note:** Version bump only for package @markuplint-dev/prettier-config
+
 ## [1.0.4](https://github.com/markuplint/markuplint/compare/@markuplint-dev/prettier-config@1.0.3...@markuplint-dev/prettier-config@1.0.4) (2024-10-14)
 
 **Note:** Version bump only for package @markuplint-dev/prettier-config