+++ packages/@markuplint/ml-config/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.8.11](https://github.com/markuplint/markuplint/compare/@markuplint/ml-config@4.8.10...@markuplint/ml-config@4.8.11) (2025-04-13)
+
+**Note:** Version bump only for package @markuplint/ml-config
+
+## [4.8.10](https://github.com/markuplint/markuplint/compare/@markuplint/ml-config@4.8.9...@markuplint/ml-config@4.8.10) (2025-03-09)
+
+**Note:** Version bump only for package @markuplint/ml-config
+
+## [4.8.9](https://github.com/markuplint/markuplint/compare/@markuplint/ml-config@4.8.8...@markuplint/ml-config@4.8.9) (2025-02-27)
+
+**Note:** Version bump only for package @markuplint/ml-config
+
+## [4.8.8](https://github.com/markuplint/markuplint/compare/@markuplint/ml-config@4.8.7...@markuplint/ml-config@4.8.8) (2025-02-11)
+
+**Note:** Version bump only for package @markuplint/ml-config
+
+## [4.8.7](https://github.com/markuplint/markuplint/compare/@markuplint/ml-config@4.8.6...@markuplint/ml-config@4.8.7) (2025-02-04)
+
+**Note:** Version bump only for package @markuplint/ml-config
+
+## [4.8.6](https://github.com/markuplint/markuplint/compare/@markuplint/ml-config@4.8.5...@markuplint/ml-config@4.8.6) (2024-12-04)
+
+**Note:** Version bump only for package @markuplint/ml-config
+
+## [4.8.5](https://github.com/markuplint/markuplint/compare/@markuplint/ml-config@4.8.4...@markuplint/ml-config@4.8.5) (2024-11-17)
+
+**Note:** Version bump only for package @markuplint/ml-config
+
+## [4.8.4](https://github.com/markuplint/markuplint/compare/@markuplint/ml-config@4.8.3...@markuplint/ml-config@4.8.4) (2024-10-31)
+
+**Note:** Version bump only for package @markuplint/ml-config
+
+## [4.8.3](https://github.com/markuplint/markuplint/compare/@markuplint/ml-config@4.8.2...@markuplint/ml-config@4.8.3) (2024-10-28)
+
+**Note:** Version bump only for package @markuplint/ml-config
+
+## [4.8.2](https://github.com/markuplint/markuplint/compare/@markuplint/ml-config@4.8.1...@markuplint/ml-config@4.8.2) (2024-10-27)
+
+**Note:** Version bump only for package @markuplint/ml-config
+
+## [4.8.1](https://github.com/markuplint/markuplint/compare/@markuplint/ml-config@4.8.0...@markuplint/ml-config@4.8.1) (2024-10-15)
+
+**Note:** Version bump only for package @markuplint/ml-config
+
 # [4.8.0](https://github.com/markuplint/markuplint/compare/@markuplint/ml-config@4.7.4...@markuplint/ml-config@4.8.0) (2024-10-14)
 
 ### Features